SLIM-1387 Refactors code and adds new class KeyHelperService
